Creating, loadouts, and leaving the map with four people "works"
Creating the team: Able to cycle through portraits but stats don't change with each click. Just need to fix the syncing there.

Selecting where to go on the map works and transitions to loadouts.

loadouts are able to be selected and filled for each player and able to start the expedition. Further testing needed here but observed all four selected people leave the map and was able to recall them back onto the map.